Added a few small details and started on resources...
Added a minimum and max damage variable. Renamed the variable
"selected_icon" to "icon_selected". Changed the digAt object to be
of type /obj/effect instead of its own object. Added in the basic
resources of stone, wood, and bone, as well as their icon file. Gave
wall turfs a 'health' variable intended to be used with digging and
deconstruction. Added the new items onto the map. Added herb, carn,
and omnivore definitions.