Inelegant Digging (Basic) Complete
Digging has now been implemented fully. There are a few issues, which
shall be further discussed in the pull request, but I don't think that
they are decisions that I should make.

Currently, I think it'll work beautifully-- So long as you only have a
single player playing. I expect a large number of fixes will need to be
done once their are multiple players, anyway, and it may be better to try
to grab them all at once.